% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asNumeral.R
\name{asNumeral}
\alias{asNumeral}
\title{Convert a number to its numeral}
\usage{
asNumeral(num)
}
\arguments{
\item{num}{any non-negative integer less than 10^27}
}
\value{
string
}
\description{
Turn a number to its English numeral.
}
\examples{
asNumeral(3945)
}
