% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build a pairty logistic model}
\usage{
buildModel(num = 1:100)
}
\arguments{
\item{num}{a vector of non-negative integers as the model's "training set"}
}
\value{
numeric
}
\description{
Returns the sigmoid midpoint of a logistic model used to predict a number's parity based on the number of vowels in its English numeral. A training set of 1 to 100 works the best.
}
\examples{
buildModel(num = 1:100)
}
