% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybeEven.R
\name{maybeEven}
\alias{maybeEven}
\title{Probability that a number is even}
\usage{
maybeEven(num, model = buildModel())
}
\arguments{
\item{num}{an integer}

\item{model}{the sigmoid midpoint from \code{buildModel()} used to predict parity}
}
\value{
numeric
}
\description{
Returns the probability from 0 to 1 that a number is even based the number of vowels in its English numeral.
}
\examples{
maybeEven(94)
}
